Rotate mask_key in read_into_buffer to fix corruption in multi-part messages.
read_into_buffer always starts reading messages using the first byte of the
mask. This is invalid for continuations of messages where the length of
any of the previous parts isn't a multiple of four.

To fix this we rotate the mask key after reading so that future messages start
reading from the next byte.